requirements.txt
pyautogui
PyWinCtl
Pillow
tqdm
loguru
customtkinter
pyrunnable